refactor(status): simplify Running to hopr_status only, move capacity/ideal_balance into BalanceResponse
